Massive Clean up

+++ Common/src/main/java/potionstudios/byg/common/world/math/noise/fastnoise/lite/FastNoiseLite.java
@@ -316,7 +316,8 @@ public class FastNoiseLite
             case OpenSimplex2:
             case OpenSimplex2S:
                 {
-                    final /*FMLdouble*/ double SQRT3 = (/*FMLdouble*/ double)1.7320508075688772935274463415059;
+                    /*FMLdouble*/
+                    final /*FMLdouble*/ double SQRT3 = 1.7320508075688772935274463415059;
                     final /*FMLdouble*/ double F2 = 0.5f * (SQRT3 - 1);
                     /*FMLdouble*/ double t = (x + y) * F2;
                     x += t;
@@ -358,25 +359,30 @@ public class FastNoiseLite
                 {
                     /*FMLdouble*/ double xy = x + y;
                     /*FMLdouble*/ double s2 = xy * -(/*FMLdouble*/ double)0.211324865405187;
-                    z *= (/*FMLdouble*/ double)0.577350269189626;
+                    /*FMLdouble*/
+                    z *= 0.577350269189626;
                     x += s2 - z;
                     y = y + s2 - z;
-                    z += xy * (/*FMLdouble*/ double)0.577350269189626;
+                    /*FMLdouble*/
+                    z += xy * 0.577350269189626;
                 }
                 break;
             case ImproveXZPlanes:
                 {
                     /*FMLdouble*/ double xz = x + z;
                     /*FMLdouble*/ double s2 = xz * -(/*FMLdouble*/ double)0.211324865405187;
-                    y *= (/*FMLdouble*/ double)0.577350269189626;
+                    /*FMLdouble*/
+                    y *= 0.577350269189626;
                     x += s2 - y;
                     z += s2 - y;
-                    y += xz * (/*FMLdouble*/ double)0.577350269189626;
+                    /*FMLdouble*/
+                    y += xz * 0.577350269189626;
                 }
                 break;
             case DefaultOpenSimplex2:
                 {
-                    final /*FMLdouble*/ double R3 = (/*FMLdouble*/ double)(2.0 / 3.0);
+                    /*FMLdouble*/
+                    final /*FMLdouble*/ double R3 = 2.0 / 3.0;
                     /*FMLdouble*/ double r = (x + y + z) * R3; // Rotation, not skew
                     x = r - x;
                     y = r - y;
@@ -935,8 +941,8 @@ public class FastNoiseLite
         float yi = (float)(y - j);
 
         float t = (xi + yi) * G2;
-        float x0 = (float)(xi - t);
-        float y0 = (float)(yi - t);
+        float x0 = xi - t;
+        float y0 = yi - t;
 
         i *= PrimeX;
         j *= PrimeY;
@@ -950,19 +956,19 @@ public class FastNoiseLite
             n0 = (a * a) * (a * a) * GradCoord(seed, i, j, x0, y0);
         }
 
-        float c = (float)(2 * (1 - 2 * G2) * (1 / G2 - 2)) * t + ((float)(-2 * (1 - 2 * G2) * (1 - 2 * G2)) + a);
+        float c = 2 * (1 - 2 * G2) * (1 / G2 - 2) * t + ((-2 * (1 - 2 * G2) * (1 - 2 * G2)) + a);
         if (c <= 0) n2 = 0;
         else
         {
-            float x2 = x0 + (2 * (float)G2 - 1);
-            float y2 = y0 + (2 * (float)G2 - 1);
+            float x2 = x0 + (2 * G2 - 1);
+            float y2 = y0 + (2 * G2 - 1);
             n2 = (c * c) * (c * c) * GradCoord(seed, i + PrimeX, j + PrimeY, x2, y2);
         }
 
         if (y0 > x0)
         {
-            float x1 = x0 + (float)G2;
-            float y1 = y0 + ((float)G2 - 1);
+            float x1 = x0 + G2;
+            float y1 = y0 + (G2 - 1);
             float b = 0.5f - x1 * x1 - y1 * y1;
             if (b <= 0) n1 = 0;
             else
@@ -972,8 +978,8 @@ public class FastNoiseLite
         }
         else
         {
-            float x1 = x0 + ((float)G2 - 1);
-            float y1 = y0 + (float)G2;
+            float x1 = x0 + (G2 - 1);
+            float y1 = y0 + G2;
             float b = 0.5f - x1 * x1 - y1 * y1;
             if (b <= 0) n1 = 0;
             else
@@ -1086,7 +1092,8 @@ public class FastNoiseLite
     {
         // 2D OpenSimplex2S case is a modified 2D simplex noise.
 
-        final /*FMLdouble*/ double SQRT3 = (/*FMLdouble*/ double)1.7320508075688772935274463415059;
+        /*FMLdouble*/
+        final /*FMLdouble*/ double SQRT3 = 1.7320508075688772935274463415059;
         final /*FMLdouble*/ double G2 = (3 - SQRT3) / 6;
 
         /*
@@ -1937,7 +1944,8 @@ public class FastNoiseLite
             case OpenSimplex2:
             case OpenSimplex2Reduced:
                 {
-                    final /*FMLdouble*/ double SQRT3 = (/*FMLdouble*/ double)1.7320508075688772935274463415059;
+                    /*FMLdouble*/
+                    final /*FMLdouble*/ double SQRT3 = 1.7320508075688772935274463415059;
                     final /*FMLdouble*/ double F2 = 0.5f * (SQRT3 - 1);
                     /*FMLdouble*/ double t = (xs + ys) * F2;
                     xs += t; ys += t;
@@ -1965,24 +1973,29 @@ public class FastNoiseLite
                 {
                     /*FMLdouble*/ double xy = xs + ys;
                     /*FMLdouble*/ double s2 = xy * -(/*FMLdouble*/ double)0.211324865405187;
-                    zs *= (/*FMLdouble*/ double)0.577350269189626;
+                    /*FMLdouble*/
+                    zs *= 0.577350269189626;
                     xs += s2 - zs;
                     ys = ys + s2 - zs;
-                    zs += xy * (/*FMLdouble*/ double)0.577350269189626;
+                    /*FMLdouble*/
+                    zs += xy * 0.577350269189626;
                 }
                 break;
             case ImproveXZPlanes:
                 {
                     /*FMLdouble*/ double xz = xs + zs;
                     /*FMLdouble*/ double s2 = xz * -(/*FMLdouble*/ double)0.211324865405187;
-                    ys *= (/*FMLdouble*/ double)0.577350269189626;
+                    /*FMLdouble*/
+                    ys *= 0.577350269189626;
                     xs += s2 - ys; zs += s2 - ys;
-                    ys += xz * (/*FMLdouble*/ double)0.577350269189626;
+                    /*FMLdouble*/
+                    ys += xz * 0.577350269189626;
                 }
                 break;
             case DefaultOpenSimplex2:
                 {
-                    final /*FMLdouble*/ double R3 = (/*FMLdouble*/ double)(2.0 / 3.0);
+                    /*FMLdouble*/
+                    final /*FMLdouble*/ double R3 = 2.0 / 3.0;
                     /*FMLdouble*/ double r = (xs + ys + zs) * R3; // Rotation, not skew
                     xs = r - xs;
                     ys = r - ys;
@@ -2014,7 +2027,8 @@ public class FastNoiseLite
                 case OpenSimplex2:
                 case OpenSimplex2Reduced:
                     {
-                        final /*FMLdouble*/ double SQRT3 = (/*FMLdouble*/ double)1.7320508075688772935274463415059;
+                        /*FMLdouble*/
+                        final /*FMLdouble*/ double SQRT3 = 1.7320508075688772935274463415059;
                         final /*FMLdouble*/ double F2 = 0.5f * (SQRT3 - 1);
                         /*FMLdouble*/ double t = (xs + ys) * F2;
                         xs += t; ys += t;
@@ -2049,24 +2063,29 @@ public class FastNoiseLite
                     {
                         /*FMLdouble*/ double xy = xs + ys;
                         /*FMLdouble*/ double s2 = xy * -(/*FMLdouble*/ double)0.211324865405187;
-                        zs *= (/*FMLdouble*/ double)0.577350269189626;
+                        /*FMLdouble*/
+                        zs *= 0.577350269189626;
                         xs += s2 - zs;
                         ys = ys + s2 - zs;
-                        zs += xy * (/*FMLdouble*/ double)0.577350269189626;
+                        /*FMLdouble*/
+                        zs += xy * 0.577350269189626;
                     }
                     break;
                 case ImproveXZPlanes:
                     {
                         /*FMLdouble*/ double xz = xs + zs;
                         /*FMLdouble*/ double s2 = xz * -(/*FMLdouble*/ double)0.211324865405187;
-                        ys *= (/*FMLdouble*/ double)0.577350269189626;
+                        /*FMLdouble*/
+                        ys *= 0.577350269189626;
                         xs += s2 - ys; zs += s2 - ys;
-                        ys += xz * (/*FMLdouble*/ double)0.577350269189626;
+                        /*FMLdouble*/
+                        ys += xz * 0.577350269189626;
                     }
                     break;
                 case DefaultOpenSimplex2:
                     {
-                        final /*FMLdouble*/ double R3 = (/*FMLdouble*/ double)(2.0 / 3.0);
+                        /*FMLdouble*/
+                        final /*FMLdouble*/ double R3 = 2.0 / 3.0;
                         /*FMLdouble*/ double r = (xs + ys + zs) * R3; // Rotation, not skew
                         xs = r - xs;
                         ys = r - ys;
@@ -2096,7 +2115,8 @@ public class FastNoiseLite
             case OpenSimplex2:
             case OpenSimplex2Reduced:
                 {
-                    final /*FMLdouble*/ double SQRT3 = (/*FMLdouble*/ double)1.7320508075688772935274463415059;
+                    /*FMLdouble*/
+                    final /*FMLdouble*/ double SQRT3 = 1.7320508075688772935274463415059;
                     final /*FMLdouble*/ double F2 = 0.5f * (SQRT3 - 1);
                     /*FMLdouble*/ double t = (xs + ys) * F2;
                     xs += t; ys += t;
@@ -2131,24 +2151,29 @@ public class FastNoiseLite
                 {
                     /*FMLdouble*/ double xy = xs + ys;
                     /*FMLdouble*/ double s2 = xy * -(/*FMLdouble*/ double)0.211324865405187;
-                    zs *= (/*FMLdouble*/ double)0.577350269189626;
+                    /*FMLdouble*/
+                    zs *= 0.577350269189626;
                     xs += s2 - zs;
                     ys = ys + s2 - zs;
-                    zs += xy * (/*FMLdouble*/ double)0.577350269189626;
+                    /*FMLdouble*/
+                    zs += xy * 0.577350269189626;
                 }
                 break;
             case ImproveXZPlanes:
                 {
                     /*FMLdouble*/ double xz = xs + zs;
                     /*FMLdouble*/ double s2 = xz * -(/*FMLdouble*/ double)0.211324865405187;
-                    ys *= (/*FMLdouble*/ double)0.577350269189626;
+                    /*FMLdouble*/
+                    ys *= 0.577350269189626;
                     xs += s2 - ys; zs += s2 - ys;
-                    ys += xz * (/*FMLdouble*/ double)0.577350269189626;
+                    /*FMLdouble*/
+                    ys += xz * 0.577350269189626;
                 }
                 break;
             case DefaultOpenSimplex2:
                 {
-                    final /*FMLdouble*/ double R3 = (/*FMLdouble*/ double)(2.0 / 3.0);
+                    /*FMLdouble*/
+                    final /*FMLdouble*/ double R3 = 2.0 / 3.0;
                     /*FMLdouble*/ double r = (xs + ys + zs) * R3; // Rotation, not skew
                     xs = r - xs;
                     ys = r - ys;
@@ -2289,8 +2314,8 @@ public class FastNoiseLite
         float yi = (float)(y - j);
 
         float t = (xi + yi) * G2;
-        float x0 = (float)(xi - t);
-        float y0 = (float)(yi - t);
+        float x0 = xi - t;
+        float y0 = yi - t;
 
         i *= PrimeX;
         j *= PrimeY;
@@ -2326,11 +2351,11 @@ public class FastNoiseLite
             vy += aaaa * yo;
         }
 
-        float c = (float)(2 * (1 - 2 * G2) * (1 / G2 - 2)) * t + ((float)(-2 * (1 - 2 * G2) * (1 - 2 * G2)) + a);
+        float c = 2 * (1 - 2 * G2) * (1 / G2 - 2) * t + ((-2 * (1 - 2 * G2) * (1 - 2 * G2)) + a);
         if (c > 0)
         {
-            float x2 = x0 + (2 * (float)G2 - 1);
-            float y2 = y0 + (2 * (float)G2 - 1);
+            float x2 = x0 + (2 * G2 - 1);
+            float y2 = y0 + (2 * G2 - 1);
             float cccc = (c * c) * (c * c);
             float xo, yo;
             if (outGradOnly)
@@ -2358,8 +2383,8 @@ public class FastNoiseLite
 
         if (y0 > x0)
         {
-            float x1 = x0 + (float)G2;
-            float y1 = y0 + ((float)G2 - 1);
+            float x1 = x0 + G2;
+            float y1 = y0 + (G2 - 1);
             float b = 0.5f - x1 * x1 - y1 * y1;
             if (b > 0)
             {
@@ -2390,8 +2415,8 @@ public class FastNoiseLite
         }
         else
         {
-            float x1 = x0 + ((float)G2 - 1);
-            float y1 = y0 + (float)G2;
+            float x1 = x0 + (G2 - 1);
+            float y1 = y0 + G2;
             float b = 0.5f - x1 * x1 - y1 * y1;
             if (b > 0)
             {
